feat: improved error handling

+++ backends/dolphinbackend/include/dolphinbackend/Exceptions.h
@@ -59,7 +59,7 @@ public:
         return std::string(what()) + 
                " [Backend: " + backend_type_ + 
                ", Operation: " + operation_ + 
-               (requested_size_ > 0 ? std::format(", Requested Size: {:.2f} GB", (static_cast<int>(requested_size_) / 1e9)) : "") + "]";
+               (requested_size_ > 0 ? std::format(", Requested Size: {:.2f} GB", (static_cast<double>(requested_size_) / 1e9)) : "") + "]";
     }
 
 private:
